# src/main/java/com/KaizakiSoftwares/JobPortal/auth/AuthenticationResponse.java
package com.KaizakiSoftwares.JobPortal.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {

    private String token;
}